requirements.txt
# Web Framework
fastapi==0.95.0
uvicorn==0.21.1

# Data Handling
pandas==2.0.0
numpy==1.24.2
ccxt==3.0.75
websocket-client>=1.5.1
python-binance>=1.0.17

# Machine Learning
tensorflow>=2.12.0
scikit-learn>=1.2.2

# Telegram/Discord Bot
python-telegram-bot>=13.15
discord.py==2.2.3
aiogram==2.25.1

# Database
sqlalchemy==1.4.48
psycopg2-binary==2.9.6
pymongo>=4.3.3
redis>=4.5.4
asyncpg==0.27.0
alembic==1.10.4

# Payment Processing
stripe==5.4.0

# Dashboard & Visualization
streamlit>=1.20.0
plotly>=5.14.0
matplotlib==3.7.1
mplfinance==0.12.9b7

# Utilities
python-dotenv==1.0.0
pydantic[email]==1.10.7
requests>=2.28.2
schedule>=1.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiocron==1.8
ta-lib==0.4.26
httpx==0.24.0
tenacity==8.2.2
aiohttp==3.8.4

# Development tools
pytest==7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
